+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.19.0]
 - Added React 19 compatibility
+- Added Pixi.js integration demo
 
 ## [0.18.3]
 - `Dropzone` component for simple drag-and-drop uploads
@@ -262,6 +265,9 @@ responsive logic uses the selected `Surface` element to handle persistent margin
 ## [v0.2.1]
 ### Other
 - vibe coded
+
+[v0.19.0]: https://github.com/off-court-creations/valet/releases/tag/v0.19.0
+
 [v0.18.3]: https://github.com/off-court-creations/valet/releases/tag/v0.18.3
 [v0.18.2]: https://github.com/off-court-creations/valet/releases/tag/v0.18.2
 [v0.18.1]: https://github.com/off-court-creations/valet/releases/tag/v0.18.1